made separate files for each component
