+++ client/components/header.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Header extends React.Component {
+  render() {
+    return <h1 className="mb-3">Student Grade Table</h1>;
+  }
+}
+
+export default Header;
