--- server/public/styles.css
@@ -1,3 +1,0 @@
-.title {
-  font-size: 2rem;
-}
